Re-organize amphtml ads examples.
