+++ build-system/app.js
@@ -45,6 +45,7 @@ const upload = multer();
 const TEST_SERVER_PORT = process.env.SERVE_PORT;
 
 app.use(bodyParser.text());
+app.use(require('./routes/a4a-envelopes'));
 app.use('/amp4test', require('./amp4test').app);
 app.use('/analytics', require('./routes/analytics'));
 app.use('/list/', require('./routes/list'));
@@ -814,60 +815,6 @@ app.get('/iframe-echo-message', (req, res) => {
   );
 });
 
-// A4A envelope.
-// Examples:
-// http://localhost:8000/a4a[-3p]/examples/animations.amp.html
-// http://localhost:8000/a4a[-3p]/proxy/s/www.washingtonpost.com/amphtml/news/post-politics/wp/2016/02/21/bernie-sanders-says-lower-turnout-contributed-to-his-nevada-loss-to-hillary-clinton/
-app.use('/a4a(|-3p)/', (req, res) => {
-  const force3p = req.baseUrl.indexOf('/a4a-3p') == 0;
-  let adUrl = req.url;
-  const templatePath = '/build-system/server-a4a-template.html';
-  const urlPrefix = getUrlPrefix(req);
-  if (!adUrl.startsWith('/proxy') && urlPrefix.indexOf('//localhost') != -1) {
-    // This is a special case for testing. `localhost` URLs are transformed to
-    // `ads.localhost` to ensure that the iframe is fully x-origin.
-    adUrl = urlPrefix.replace('localhost', 'ads.localhost') + adUrl;
-  }
-  adUrl = addQueryParam(adUrl, 'inabox', 1);
-  fs.readFileAsync(pc.cwd() + templatePath, 'utf8').then(template => {
-    const result = template
-      .replace(/CHECKSIG/g, force3p || '')
-      .replace(/DISABLE3PFALLBACK/g, !force3p)
-      .replace(/OFFSET/g, req.query.offset || '0px')
-      .replace(/AD_URL/g, adUrl)
-      .replace(/AD_WIDTH/g, req.query.width || '300')
-      .replace(/AD_HEIGHT/g, req.query.height || '250');
-    res.end(result);
-  });
-});
-
-// In-a-box envelope.
-// Examples:
-// http://localhost:8000/inabox/examples/animations.amp.html
-// http://localhost:8000/inabox/proxy/s/www.washingtonpost.com/amphtml/news/post-politics/wp/2016/02/21/bernie-sanders-says-lower-turnout-contributed-to-his-nevada-loss-to-hillary-clinton/
-app.use('/inabox/', (req, res) => {
-  let adUrl = req.url;
-  const templatePath = '/build-system/server-inabox-template.html';
-  const urlPrefix = getUrlPrefix(req);
-  if (
-    !adUrl.startsWith('/proxy') && // Ignore /proxy
-    urlPrefix.indexOf('//localhost') != -1
-  ) {
-    // This is a special case for testing. `localhost` URLs are transformed to
-    // `ads.localhost` to ensure that the iframe is fully x-origin.
-    adUrl = urlPrefix.replace('localhost', 'ads.localhost') + adUrl;
-  }
-  adUrl = addQueryParam(adUrl, 'inabox', 1);
-  fs.readFileAsync(pc.cwd() + templatePath, 'utf8').then(template => {
-    const result = template
-      .replace(/AD_URL/g, adUrl)
-      .replace(/OFFSET/g, req.query.offset || '0px')
-      .replace(/AD_WIDTH/g, req.query.width || '300')
-      .replace(/AD_HEIGHT/g, req.query.height || '250');
-    res.end(result);
-  });
-});
-
 /**
  * Append ?sleep=5 to any included JS file in examples to emulate delay in
  * loading that file. This allows you to test issues with your extension being
@@ -1343,23 +1290,6 @@ function addViewerIntegrationScript(ampJsVersion, file) {
 
 function getUrlPrefix(req) {
   return req.protocol + '://' + req.headers.host;
-}
-
-/**
- * @param {string} url
- * @param {string} param
- * @param {*} value
- * @return {string}
- */
-function addQueryParam(url, param, value) {
-  const paramValue =
-    encodeURIComponent(param) + '=' + encodeURIComponent(value);
-  if (!url.includes('?')) {
-    url += '?' + paramValue;
-  } else {
-    url += '&' + paramValue;
-  }
-  return url;
 }
 
 function generateInfo(filePath) {
