+++ build-system/server-inabox-template.html
@@ -31,6 +31,6 @@
       frameborder="0">
   </iframe>
 
-  <script src="/examples/inabox-tag-integration.js"></script>
+  <script src="/examples/amphtml-ads/ads-tag-integration.js"></script>
 </body>
 </html>
