+++ examples/amphtml-ads/README.md
@@ -1,0 +1,2 @@
+The examples in this folder has to be viewed through one of the 
+supported [AMPHTML ads envelopes](../../../../contributing/TESTING.md#a4a-envelope-a4a-a4a-3p).
