+++ examples/amphtml-ads/adchoices-1.a4a.html
