+++ examples/amphtml-ads/adchoices-2.a4a.html
