+++ examples/amphtml-ads/ads-tag-integration.js
