+++ examples/amphtml-ads/fake-memory.host.html
@@ -51,7 +51,7 @@
       iframe = document.createElement('iframe');
       iframe.name = 'iframe-' + n;
       n++;
-      iframe.src = '//ads.localhost:8000/examples/inabox.amp.html';
+      iframe.src = '//ads.localhost:8000/examples/amphtml-ads/image-ad.a4a.html';
       iframe.scrolling = 'no';
       iframe.width = 300;
       iframe.height = 200;
@@ -76,5 +76,5 @@
     }
     repeatedly_replace_ad_iframe();
   </script>
-  <script src="./inabox-tag-integration.js"></script>
+  <script src="/examples/amphtml-ads/ads-tag-integration.js"></script>
 </body>
