+++ examples/amphtml-ads/gpt.host.html
@@ -13,7 +13,7 @@
 
   <h1>AMP inabox for GPT demo</h1>
   <h2>Standard GPT Ad</h2>
-  <h4><i>served in an iframe from examples/ads/inabox.adchoices.html</i></h4>
+  <h4><i>served in an iframe from /examples/amphtml-ads/adchoices-1.a4a.html</i></h4>
   <div id="div-gpt-ad-1533830709700-0" style="height:300px; width:600px;" data-google-query-id="CMDAi-P-6N0CFUuJwAoddusOag">
     <iframe id="google_ads_iframe_/49790152/scaledplacement/300x300_0"
       title="3rd party ad content"
@@ -24,13 +24,13 @@
       marginwidth="0"
       marginheight="0"
       frameborder="0"
-      src="/examples/ads/inabox.adchoices.html"
+      src="/examples/amphtml-ads/adchoices-1.a4a.html"
       style="border: 0px; vertical-align: bottom;"
       data-load-complete="true">
     </iframe>
   </div>
 
-  <h4><i>served in an iframe from examples/ads/inabox.adchoices.creative.html</i></h4>
+  <h4><i>served in an iframe from /examples/amphtml-ads/adchoices-2.a4a.html</i></h4>
   <div id="div-gpt-ad-1533830709700-0" data-google-query-id="CMDAi-P-6N0CFUuJwAoddusOag">
     <iframe id="google_ads_iframe_/49790152/scaledplacement/300x300_0"
             title="3rd party ad content"
@@ -41,7 +41,7 @@
             marginwidth="0"
             marginheight="0"
             frameborder="0"
-            src="/examples/ads/inabox.adchoices.creative.html"
+            src="/examples/amphtml-ads/adchoices-2.a4a.html"
             style="border: 0px; vertical-align: bottom;"
             data-load-complete="true">
     </iframe>
