+++ examples/visual-tests/amp-ad/amp-inabox-creative.html
@@ -21,4 +21,4 @@ window.addEventListener('message', function(event) {
       scrolling="no"
       width="300" height="200">
   </iframe>
-  <script src="/examples/inabox-tag-integration.js"></script>
+  <script src="/examples/amphtml-ads/ads-tag-integration.js"></script>
