+++ test/manual/amp-animation/scrollbound-inabox.non.amp.html
@@ -58,6 +58,6 @@
     Sed sed augue auctor, porta turpis ultrices, cursus diam. In venenatis aliquet porta. Sed volutpat fermentum quam, ac
     molestie nulla porttitor ac. Donec porta risus ut enim pellentesque, id placerat elit ornare.
   </p>
-  <script src="../../../examples/inabox-tag-integration.js"></script>
+  <script src="../../../examples/amphtml-ads/ads-tag-integration.js"></script>
 </body>
 </html>
